+++ src/lightning/pytorch/trainer/supporters.py
@@ -281,6 +281,7 @@ class CombinedLoader(Iterable):
 
     def _update_index(self, dataloader: Iterable, index: int) -> None:
         # mutation needs to be done using this method to avoid stale references
+        # TODO(carmocca): avoid this, inefficient
         self._flattened[index] = dataloader
         self._iterables = tree_unflatten(self._flattened, self._spec)