+++ src/lightning/pytorch/utilities/types.py
@@ -18,11 +18,10 @@ Convention:
 """
 from contextlib import contextmanager
 from dataclasses import dataclass
-from typing import Any, Dict, Generator, List, Optional, Protocol, runtime_checkable, Sequence, Type, Union
+from typing import Any, Dict, Generator, List, Optional, Protocol, runtime_checkable, Type, Union
 
 import torch
 from torch import Tensor
-from torch.utils.data import DataLoader
 from torchmetrics import Metric
 
 from lightning.fabric.utilities.types import _TORCH_LRSCHEDULER, LRScheduler, ProcessGroup, ReduceLROnPlateau
@@ -32,16 +31,8 @@ _METRIC = Union[Metric, Tensor, _NUMBER]
 STEP_OUTPUT = Union[Tensor, Dict[str, Any]]
 _EVALUATE_OUTPUT = List[Dict[str, float]]  # 1 dict per DataLoader
 _PREDICT_OUTPUT = Union[List[Any], List[List[Any]]]
-TRAIN_DATALOADERS = Union[
-    DataLoader,
-    Sequence[DataLoader],
-    Sequence[Sequence[DataLoader]],
-    Sequence[Dict[str, DataLoader]],
-    Dict[str, DataLoader],
-    Dict[str, Dict[str, DataLoader]],
-    Dict[str, Sequence[DataLoader]],
-]
-EVAL_DATALOADERS = Union[DataLoader, Sequence[DataLoader]]
+TRAIN_DATALOADERS = Any  # any iterable or collection of iterables
+EVAL_DATALOADERS = Any  # any iterable or collection of iterables
 
 
 @runtime_checkable