+++ tests/tests_pytorch/strategies/test_xla.py
@@ -21,33 +21,16 @@ from torch.utils.data import DataLoader
 
 from lightning.pytorch import Trainer
 from lightning.pytorch.demos.boring_classes import BoringModel, RandomDataset
-from lightning.pytorch.strategies import TPUSpawnStrategy
+from lightning.pytorch.strategies import XLAStrategy
 from tests_pytorch.helpers.dataloaders import CustomNotImplementedErrorDataloader
 from tests_pytorch.helpers.runif import RunIf
 
 
-class BoringModelNoDataloaders(BoringModel):
-    def train_dataloader(self):
-        raise NotImplementedError
-
-    def val_dataloader(self):
-        raise NotImplementedError
-
-    def test_dataloader(self):
-        raise NotImplementedError
-
-    def predict_dataloader(self):
-        raise NotImplementedError
-
-
-_loader = DataLoader(RandomDataset(32, 64))
-_loader_no_len = CustomNotImplementedErrorDataloader(_loader)
-
-
 def test_error_process_iterable_dataloader(xla_available):
-    strategy = TPUSpawnStrategy(MagicMock())
+    strategy = XLAStrategy(MagicMock())
+    loader_no_len = CustomNotImplementedErrorDataloader(DataLoader(RandomDataset(32, 64)))
     with pytest.raises(TypeError, match="TPUs do not currently support"):
-        strategy.process_dataloader(_loader_no_len)
+        strategy.process_dataloader(loader_no_len)
 
 
 class BoringModelTPU(BoringModel):
@@ -60,9 +43,9 @@ class BoringModelTPU(BoringModel):
 @RunIf(tpu=True, standalone=True)
 @mock.patch.dict(os.environ, os.environ.copy(), clear=True)
 def test_model_tpu_one_core():
-    """Tests if device/debug flag is set correctly when training and after teardown for TPUSpawnStrategy."""
+    """Tests if device/debug flag is set correctly when training and after teardown for XLAStrategy."""
     model = BoringModelTPU()
-    trainer = Trainer(accelerator="tpu", devices=1, fast_dev_run=True, strategy=TPUSpawnStrategy(debug=True))
-    assert isinstance(trainer.strategy, TPUSpawnStrategy)
+    trainer = Trainer(accelerator="tpu", devices=1, fast_dev_run=True, strategy=XLAStrategy(debug=True))
+    assert isinstance(trainer.strategy, XLAStrategy)
     trainer.fit(model)
     assert "PT_XLA_DEBUG" not in os.environ